Bugs fixes for Trello 174/184. These are the ones from Jira
