+++ patientview-parent/radar/src/test/java/org/patientview/radar/test/util/RadarUtilityTest.java
@@ -4,6 +4,7 @@ package org.patientview.radar.test.util;
 import org.junit.Assert;
 import org.junit.Test;
 import org.patientview.radar.util.RadarUtility;
+import org.patientview.util.CommonUtils;
 
 import java.util.Calendar;
 import java.util.Date;
@@ -33,7 +34,7 @@ public class RadarUtilityTest {
         final String testDateString2 = "28-11-95";
         final String testDateString3 = "1964-05-12";
 
-        Date date = RadarUtility.parseDate(testDateString1);
+        Date date = CommonUtils.parseDate(testDateString1);
 
         Calendar calendar = Calendar.getInstance();
         calendar.setTime(date);
@@ -41,14 +42,14 @@ public class RadarUtilityTest {
         Assert.assertTrue("The month for " + testDateString1 + " should 2", calendar.get(Calendar.MONTH) == 1);
         Assert.assertTrue("The day of the month for " + testDateString1 + " should 2", calendar.get(Calendar.DAY_OF_MONTH) == 8);
 
-        date = RadarUtility.parseDate(testDateString2);
+        date = CommonUtils.parseDate(testDateString2);
 
         calendar.setTime(date);
         Assert.assertTrue("The year for " + testDateString2 + " should 1995", calendar.get(Calendar.YEAR) == 1995);
         Assert.assertTrue("The month for " + testDateString2 + " should 11", calendar.get(Calendar.MONTH) == 10);
         Assert.assertTrue("The day of the month for " + testDateString2 + " should 28", calendar.get(Calendar.DAY_OF_MONTH) == 28);
 
-        date = RadarUtility.parseDate(testDateString3);
+        date = CommonUtils.parseDate(testDateString3);
 
         calendar.setTime(date);
         Assert.assertTrue("The year for " + testDateString3 + " should 1964", calendar.get(Calendar.YEAR) == 1964);
